backend/requirements.txt
fastapi
uvicorn
playwright
google-cloud-aiplatform
google-genai
python-dotenv
pydantic
reportlab
aiofiles
httpx
jinja2
python-multipart
svglib
playwright-stealth
pillow
cffi
